more changes to unsigned ints

+++ gpu.cu
@@ -58,7 +58,7 @@ __device__ void calcArrayIndex(int *index, int adjN, int adjY, int x){
 }
 
 // Calculate the position in the matrix
-__device__ void calcPosInMatrix(int index, int n, int *x, int *y){
+__device__ void calcPosInMatrix(int index, int n, unsigned short *x, unsigned short *y){
   calcXPos(x, index * 2, n - (.5f));
   calcYPos(y, index + 1, 3 - 2 * n, *x);
 }
